+++ src/js/PlotBoilerplate.js
@@ -1434,19 +1434,4 @@ var PlotBoilerplate = /** @class */ (function () {
     return PlotBoilerplate;
 }()); // END class PlotBoilerplate
 exports.PlotBoilerplate = PlotBoilerplate;
-// const test : PlotBoilerplate = new PlotBoilerplate( ({} as unknown) as Config );
-//exposeClass(PlotBoilerplate);
-//getGlobal()["PlotBoilerplate"] = PlotBoilerplate;
-// NOT WORKING
-// console.log( "exposeClass", classObject.constructor.name );
-/* ((_context:any) => {
-    // console.log( "exposeClass", classObject, classObject.constructor );
-    // const exposeClass : string = "TEST";
-    console.log('Expose plotboilerplate');
-    _context.PlotBoilerplate = PlotBoilerplate;
-})(typeof window !== 'undefined' ? window : this);
-*/
-// ((window || globalThis) as any)["PlotBoilerplate"] = PlotBoilerplate;
-// console.log('test');
-// (window as any).PlotBoilerplate = PlotBoilerplate;
 //# sourceMappingURL=PlotBoilerplate.js.map