+++ src/js/Triangle.js
@@ -21,7 +21,10 @@
  * @modified  2020-03-17 Added proper JSDoc comments.
  * @modified  2020-03-25 Ported this class from vanilla-JS to Typescript.
  * @modified  2020-05-09 Added the new Circle class (ported to Typescript from the demos).
- * @version   2.2.3
+ * @modified  2020-05-12 Added getIncircularTriangle() function.
+ * @modified  2020-05-12 Added getIncircle() function.
+ * @modified  2020-05-12 Fixed the signature of getCircumcirle(). Was still a generic object.
+ * @version   2.2.4
  *
  * @file Triangle
  * @fileoverview A simple triangle class: three vertices.
@@ -30,8 +33,10 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var Bounds_1 = require("./Bounds");
 var Circle_1 = require("./Circle");
+var Line_1 = require("./Line");
 var Polygon_1 = require("./Polygon");
 var Vertex_1 = require("./Vertex");
+var geomutils_1 = require("./geomutils");
 var Triangle = /** @class */ (function () {
     /**
      * The constructor.
@@ -274,6 +279,37 @@ var Triangle = /** @class */ (function () {
      */
     Triangle.prototype.containsPoint = function (p) {
         return Triangle.utils.pointIsInTriangle(p.x, p.y, this.a.x, this.a.y, this.b.x, this.b.y, this.c.x, this.c.y);
+    };
+    ;
+    /**
+     * Get that inner triangle which defines the maximal incircle.
+     *
+     * @return {Triangle} The triangle of those points in this triangle that define the incircle.
+     */
+    Triangle.prototype.getIncircularTriangle = function () {
+        var lineA = new Line_1.Line(this.a, this.b);
+        var lineB = new Line_1.Line(this.b, this.c);
+        var lineC = new Line_1.Line(this.c, this.a);
+        var bisector1 = geomutils_1.geomutils.nsectAngle(this.b, this.a, this.c, 2)[0]; // bisector of first angle (in b)
+        var bisector2 = geomutils_1.geomutils.nsectAngle(this.c, this.b, this.a, 2)[0]; // bisector of second angle (in c)
+        var intersection = bisector1.intersection(bisector2);
+        // Find the closest points on one of the polygon lines (all have same distance by construction)
+        var circleIntersA = lineA.getClosestPoint(intersection);
+        var circleIntersB = lineB.getClosestPoint(intersection);
+        var circleIntersC = lineC.getClosestPoint(intersection);
+        return new Triangle(circleIntersA, circleIntersB, circleIntersC);
+    };
+    ;
+    /**
+     * Get the incircle of this triangle. That is the circle that touches each side
+     * of this triangle in exactly one point.
+     *
+     * Note this just calls getIncircularTriangle().getCircumcircle()
+     *
+     * @return {Circle} The incircle of this triangle.
+     */
+    Triangle.prototype.getIncircle = function () {
+        return this.getIncircularTriangle().getCircumcircle();
     };
     ;
     /**