+++ src/js/VertTuple.js
@@ -5,7 +5,10 @@
  * @requires Vertex
  *
  * @author Ikaros Kappler
+ * @date   2020-03-24
  * @modified 2020-05-04 Fixed a serious bug in the pointDistance function.
+ * @modofied 2020-05-12 The angle(line) param was still not optional. Changed that.
+ * @version 1.0.1
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 var Vertex_1 = require("./Vertex");
@@ -124,14 +127,14 @@ var VertTuple = /** @class */ (function () {
      * Get the angle between this and the passed line (in radians).
      *
      * @method angle
-     * @param {VertTuple} [line] - (optional) The line to calculate the angle to. If null the baseline (x-axis) will be used.
+     * @param {VertTuple} line - (optional) The line to calculate the angle to. If null the baseline (x-axis) will be used.
      * @return {number} this
      * @instance
      * @memberof VertTuple
      **/
     VertTuple.prototype.angle = function (line) {
         if (typeof line == 'undefined')
-            line = this.factory(new Vertex_1.Vertex(0, 0), new Vertex_1.Vertex(100, 0)); // new Line( new Vertex(0,0), new Vertex(100,0) );
+            line = this.factory(new Vertex_1.Vertex(0, 0), new Vertex_1.Vertex(100, 0));
         // Compute the angle from x axis and the return the difference :)
         var v0 = this.b.clone().sub(this.a);
         var v1 = line.b.clone().sub(line.a);