+++ src/ts/VertTuple.ts
@@ -4,7 +4,10 @@
  * @requires Vertex 
  *
  * @author Ikaros Kappler
+ * @date   2020-03-24
  * @modified 2020-05-04 Fixed a serious bug in the pointDistance function.
+ * @modofied 2020-05-12 The angle(line) param was still not optional. Changed that.
+ * @version 1.0.1
  */
 
 import { Vertex } from "./Vertex";
@@ -155,14 +158,14 @@ export class VertTuple<T extends VertTuple<T>> {
      * Get the angle between this and the passed line (in radians).
      *
      * @method angle
-     * @param {VertTuple} [line] - (optional) The line to calculate the angle to. If null the baseline (x-axis) will be used.
+     * @param {VertTuple} line - (optional) The line to calculate the angle to. If null the baseline (x-axis) will be used.
      * @return {number} this
      * @instance
      * @memberof VertTuple
      **/
-    angle( line:VertTuple<any> ):number {
+    angle( line?:VertTuple<any> ):number {
 	if( typeof line == 'undefined' )
-	    line = this.factory( new Vertex(0,0), new Vertex(100,0) ); // new Line( new Vertex(0,0), new Vertex(100,0) );
+	    line = this.factory( new Vertex(0,0), new Vertex(100,0) );
 	// Compute the angle from x axis and the return the difference :)
 	var v0 = this.b.clone().sub( this.a );
 	var v1 = line.b.clone().sub( line.a );